Adds LocalUserDataStateService protocol & default implementation. Also adds LocalUserDataKeyAppSettingsStore.
